Initialize od_dec_blank_img correctly
It appears od_dec_blank_img was missed in the recent 2X upsampled reference -> 1X reference
changes. Altered initialization to 1X frame size.

In addition, it was sizing the initialization including borders, but did not offset the
start of initialization to point to the beginning of the buffer instead of the beginning
of the visible area.

Note this code is excercized only during decode when beginning on a non-keyframe.