Move MV vis code to encode.c
The MV visualization code is now entirely encode-specific-- and it
relies on buffers that will soon exist only on encode side in the
od_enc_ctx.  Move the relevant routines to encode.c

I get that many of these drawing routines were (or were intended to
be) general purpose and called from decode as well, but that's unlikely to
ever be the case again in the future given the new analyzers.