Eliminate previous use of xstride for upsampling
Previously, 8-bit planes could be upsample, and this was signalled
with an xstride of 2.  Upsampling is no longer used, but >8-bit frames
now have an xstride of 2.

This updates the convention such that 8- and 16- bit specific
functions are named accordingly, and do not take an xstride, as it is
known at the point we are branching between 8 and 16 bit.