Centralize most encoder lambda calculation
The encoder uses many, many lambda values, most of which are calculated directly
or indirectly from the quantizer value in use.  Put them in one place, in the
same place that will be handling quantizer determination.

The exception to this is the motion compensation lambda; this is calculated in
the motion estimation initialization code, and I've left it alone.